Squashed commit of the following:
commit 35d5446b946f7695985b88b4a791d3d6c301b39a
Author: volkerschulz <github@volkerschulz.de>
Date:   Mon Oct 30 08:58:27 2023 +0100

    Update README.md

commit 29ea8b46e145ac2d65e0b40a4e0d699fdac1da20
Author: Volker Schulz <volker.schulz@deinerstertag.de>
Date:   Mon Oct 30 08:34:42 2023 +0100

    Create CHANGELOG.md

commit d8ee3ba3c8fd56fad6a5fd8a51734f742ce23da7
Author: Volker Schulz <volker.schulz@deinerstertag.de>
Date:   Mon Oct 30 08:33:36 2023 +0100

    Create screenshot-1.png

commit 0e17056a1bee228ee27bc64f35ddfa2e1fcdce02
Author: Volker Schulz <volker.schulz@deinerstertag.de>
Date:   Mon Oct 30 08:13:30 2023 +0100

    Check for file before unlinking

commit c15226974114e6959155c324d6541493e1f086a7
Author: volkerschulz <github@volkerschulz.de>
Date:   Mon Oct 30 03:44:46 2023 +0100

    Update README.md

commit 0ac156a33499485aa7ee64dfd343427b5f0269f7
Merge: 7a000e3 af5f86b
Author: Volker Schulz <volker.schulz@deinerstertag.de>
Date:   Sun Oct 29 20:16:58 2023 +0100

    Merge branch 'develop' of https://github.com/volkerschulz/WordPress-Error-Reporting into develop

commit 7a000e3723ad917ed041e507f38b059e96841e20
Author: Volker Schulz <volker.schulz@deinerstertag.de>
Date:   Sun Oct 29 20:16:56 2023 +0100

    Update class-error-reporting-deactivator.php

commit 0e7b53ddcaf965d87f74e53c1f73c18157e2ed8f
Author: Volker Schulz <volker.schulz@deinerstertag.de>
Date:   Sun Oct 29 20:14:45 2023 +0100

    Update class-error-reporting-deactivator.php

+++ error-reporting/includes/class-error-reporting-deactivator.php
@@ -31,10 +31,10 @@ class Error_Reporting_Deactivator {
 	public static function deactivate() {
 		$target_file = get_option('error-reporting-mu-file');
 		if(!empty($target_file)) {
-			if(!unlink($target_file)) {
+			if(file_exists($target_file) && !unlink($target_file)) {
 				trigger_error(
 					'Could not unlink mu-file', 
-					E_USER_ERROR
+					E_USER_WARNING
 				);
 			}
 			delete_option('error-reporting-mu-file');
